Move request parsers into a separate file
